Split out `ModuleContext`
