+++ lib/loader.js
@@ -5,7 +5,7 @@ const { isURL, pathToFileURL } = require('bare-url')
 const { urlToPath } = require('./url')
 const ModuleProtocol = require('./protocol')
 const ModuleSource = require('./source')
-const { setDefaultLoader } = require('./module')
+const ModuleContext = require('./context')
 const SourceTextModule = require('./module/source-text')
 const ScriptModule = require('./module/script')
 const JSONModule = require('./module/json')
@@ -85,7 +85,7 @@ module.exports = class ModuleLoader {
     this._uninstantiated = []
     this._pending = null
 
-    setDefaultLoader(this)
+    ModuleContext.setDefaultLoader(this)
   }
 
   get protocol() {
