+++ lib/module.js
@@ -1,32 +1,21 @@
 const lex = require('bare-module-lexer')
-const { pathToFileURL } = require('bare-url')
 const strip = require('bare-type-stripper')
 const binding = require('../binding')
 const { urlToPath, urlToDirname } = require('./url')
-const errors = require('./errors')
+const ModuleContext = require('./context')
 
 const { REEXPORT, ADDON, ASSET } = lex.constants
 
-const status = {
-  UNLINKED: 0,
-  LINKED: 1,
-  EVALUATING: 2,
-  EVALUATED: 3
-}
-
-const records = new Map()
-
-let defaultLoader = null
-
-function setDefaultLoader(loader) {
-  defaultLoader = loader
-}
+const UNLINKED = 0
+const LINKED = 1
+const EVALUATING = 2
+const EVALUATED = 3
 
 module.exports = exports = class Module {
   constructor(loader, source) {
     this._loader = loader
     this._source = source
-    this._status = status.UNLINKED
+    this._status = UNLINKED
 
     this._id = null
     this._names = null
@@ -132,21 +121,21 @@ module.exports = exports = class Module {
   }
 
   _instantiate() {
-    if (this._status >= status.LINKED) return
+    if (this._status >= LINKED) return
 
     this._initialize()
 
-    this._status = status.LINKED
+    this._status = LINKED
   }
 
   _evaluate() {
     if (this._error !== null) throw this._error
 
-    if (this._status >= status.EVALUATING) return this.exports
+    if (this._status >= EVALUATING) return this.exports
 
     this._instantiate()
 
-    this._status = status.EVALUATING
+    this._status = EVALUATING
 
     try {
       this._execute()
@@ -156,7 +145,7 @@ module.exports = exports = class Module {
       throw err
     }
 
-    this._status = status.EVALUATED
+    this._status = EVALUATED
 
     return this.exports
   }
@@ -180,19 +169,19 @@ module.exports = exports = class Module {
   }
 
   _text() {
-    return stripTypeScript(this._source.bytes, extname(this.url)).toString()
+    return stripTypeScript(this._source.bytes, this.url).toString()
   }
 
   _createModule(source) {
-    this._id = binding.createModule(Module, this, this.url.href, source, 0)
+    this._id = binding.createModule(ModuleContext, this, this.url.href, source, 0)
 
-    records.set(this._id, this)
+    ModuleContext.add(this)
   }
 
   _createSyntheticModule(names) {
-    this._id = binding.createSyntheticModule(Module, this, this.url.href, names)
+    this._id = binding.createSyntheticModule(ModuleContext, this, this.url.href, names)
 
-    records.set(this._id, this)
+    ModuleContext.add(this)
   }
 
   _createFunction(source) {
@@ -203,145 +192,27 @@ module.exports = exports = class Module {
       0
     )
 
-    records.set(binding.getFunctionID(fn), this)
+    this._id = binding.getFunctionID(fn)
+
+    ModuleContext.add(this)
 
     return fn
   }
 
   _runModule() {
-    return binding.runModule(Module, this, onrun)
+    return binding.runModule(ModuleContext, this, this._onrun)
   }
 
   _moduleNamespace() {
     return binding.getModuleNamespace(this)
   }
-}
 
-binding.init(exports, onimport, onevaluate, onmeta)
+  _onrun(reason, promise, err = reason) {
+    if (err) {
+      promise.catch(() => {}) // Don't leak the rejection before it is rethrown.
 
-exports.setDefaultLoader = setDefaultLoader
-
-function onimport(specifier, attributes, referrerName, id, isDynamicImport) {
-  const referrer = records.get(id) || null
-
-  if (isDynamicImport) {
-    return ondynamicimport(referrer, specifier, referrerName)
-  }
-
-  if (referrer === null) {
-    throw errors.MODULE_NOT_FOUND(`Cannot find referrer for module '${specifier}'`, specifier)
-  }
-
-  const url = referrer._resolve(specifier, 'import')
-
-  const child = referrer._loader._lookup(url)
-
-  if (!child) {
-    throw errors.MODULE_NOT_FOUND(
-      `Cannot find module '${specifier}' imported from '${referrer.url.href}'`,
-      specifier,
-      referrer.url
-    )
-  }
-
-  child._instantiate()
-  child._synthesize()
-
-  return child
-}
-
-async function ondynamicimport(referrer, specifier, referrerName) {
-  const loader = referrer !== null ? referrer._loader : defaultLoader
-
-  if (loader === null) {
-    throw errors.MODULE_NOT_FOUND(
-      `Cannot find referrer for dynamically imported module '${specifier}'`,
-      specifier
-    )
-  }
-
-  let url = referrer !== null ? referrer._resolve(specifier, 'import') : null
-
-  let record = loader._lookup(url)
-
-  if (record === null) {
-    const parentURL = referrer !== null ? referrer.url : referrerURL(referrerName)
-
-    url = await loader._resolveAsync(specifier, parentURL, 'import')
-
-    record = await loader.link(url)
-  } else {
-    record._instantiate()
-  }
-
-  record._synthesize()
-
-  await record._run()
-
-  return record._moduleNamespace()
-}
-
-function onevaluate(id) {
-  const record = records.get(id)
-
-  record._evaluate()
-
-  for (const name of record._names) {
-    let value
-
-    if (
-      name === 'default' &&
-      (typeof record.exports !== 'object' ||
-        record.exports === null ||
-        name in record.exports === false)
-    ) {
-      value = record.exports
-    } else {
-      value = record.exports[name]
+      throw err
     }
-
-    binding.setModuleExport(record, name, value)
-  }
-}
-
-function onmeta(id, meta) {
-  const record = records.get(id)
-
-  meta.url = record.url.href
-  meta.main = record._loader.main === record
-
-  // For Node.js compatibility.
-  meta.dirname = record.dirname
-  meta.filename = record.filename
-
-  meta.resolve = function (specifier) {
-    const url = record._resolve(specifier, 'import')
-
-    return (url || record._loader._resolveSync(specifier, record.url, 'import')).href
-  }
-
-  meta.asset = function (specifier) {
-    return record._loader._resolveArtifact(record, specifier, 'asset').href
-  }
-
-  meta.addon = function (specifier = '.') {
-    const url = record._loader._resolveArtifact(record, specifier, 'addon')
-
-    return Bare.Addon.load(url, { referrer: record }).exports
-  }
-
-  meta.addon.resolve = function (specifier = '.') {
-    return record._loader._resolveArtifact(record, specifier, 'addon').href
-  }
-
-  meta.addon.host = Bare.Addon.host
-}
-
-function onrun(reason, promise, err = reason) {
-  if (err) {
-    promise.catch(() => {}) // Don't leak the rejection before it is rethrown.
-
-    throw err
   }
 }
 
@@ -355,20 +226,8 @@ function pickCondition(entry, condition) {
   return null
 }
 
-function referrerURL(referrerName) {
-  if (!referrerName) return pathToFileURL('./')
+const typeScriptExtension = /\.(c|m)?ts$/
 
-  return URL.parse(referrerName) || pathToFileURL(referrerName)
-}
-
-function extname(url) {
-  const match = url.pathname.match(/\.[^./]+$/)
-
-  return match === null ? '' : match[0]
-}
-
-const typeScriptExtensions = new Set(['.ts', '.cts', '.mts'])
-
-function stripTypeScript(source, extension) {
-  return typeScriptExtensions.has(extension) ? strip(source) : source
+function stripTypeScript(source, url) {
+  return typeScriptExtension.test(url.pathname) ? strip(source) : source
 }
