+++ lib/module/synthetic.js
@@ -1,9 +1,16 @@
 const Module = require('../module')
 
 module.exports = class SyntheticModule extends Module {
-  _synthesize() {
-    if (this._id !== null) return
+  constructor(loader, source) {
+    super(loader, source)
 
+    this._synthesized = false
+  }
+
+  _synthesize() {
+    if (this._synthesized) return
+
+    this._synthesized = true
     this._names = Array.from(this._exportNames())
 
     this._createSyntheticModule(this._names)
